chore: refactore rate us button with modal to a single component, return a splash screen on load
