--- src/Components/SupportUsForm/types.ts
@@ -1,3 +1,0 @@
-export type SupportUsFormPropTypes = {
-  onComplete?: () => void;
-};
